requirements.txt
fastapi[standard]
asyncpg
pydantic-settings # to manage env variables in fastapi
sqlmodel
alembic
passlib
